move routeinfo and stationInfo to separate files
